Diseño mejorado para medicamentos, ajustes en ViewPager2 y botón agregado
