+++ app/build.gradle
@@ -76,4 +76,7 @@ dependencies {
 
     // leanback-tab is an add-on for a top navigation bar.
     implementation "androidx.leanback:leanback-tab:1.1.0-beta01"
+    implementation 'com.google.android.material:material:1.12.0'
+    implementation libs.androidx.constraintlayout
+    implementation libs.androidx.constraintlayout.v214
 }
