+++ app/src/main/res/layout/activity_register.xml
@@ -57,7 +57,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:text="Registrar"
-        android:backgroundTint="@color/primary_button"
+        android:backgroundTint="@color/button_primary"
         android:textColor="@color/white"
         android:layout_marginBottom="16dp" />
