+++ app/src/main/res/values-sw540dp/dimens.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <!-- Margen para la Guideline -->
+    <dimen name="guideline_percent">0.05</dimen>
+
+    <!-- Tamaño de texto para el título -->
+    <dimen name="title_text_size">32sp</dimen>
+
+    <!-- Dimensiones de los botones -->
+    <dimen name="button_width">400dp</dimen>
+    <dimen name="button_height">70dp</dimen>
+
+    <!-- Tamaño de texto de los botones -->
+    <dimen name="button_text_size">18sp</dimen>
+
+    <!-- Margen superior de los botones -->
+    <dimen name="button_margin_top">20dp</dimen>
+</resources>
