+++ app/src/main/res/values-sw600dp/dimens.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <dimen name="guideline_percent">0.12</dimen>
+    <dimen name="title_text_size">32sp</dimen>
+    <dimen name="button_width">400dp</dimen>
+    <dimen name="button_height">80dp</dimen>
+    <dimen name="button_text_size">18sp</dimen>
+    <dimen name="button_margin_top">30dp</dimen>
+</resources>
