+++ app/src/main/res/values-sw720dp/dimens.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <dimen name="guideline_percent">0.10</dimen>
+    <dimen name="title_text_size">36sp</dimen>
+    <dimen name="button_width">500dp</dimen>
+    <dimen name="button_height">90dp</dimen>
+    <dimen name="button_text_size">20sp</dimen>
+    <dimen name="button_margin_top">40dp</dimen>
+</resources>
