+++ app/src/main/res/values-v28/style.xml
@@ -23,4 +23,8 @@
         <item name="android:textColor">@color/primary_text</item>
     </style>
 
+    <style name="ShapeAppearanceOverlay" parent="ShapeAppearance.MaterialComponents.SmallComponent">
+        <item name="cornerFamily">rounded</item>
+        <item name="cornerSize">8dp</item>
+    </style>
 </resources>
