+++ app/src/main/res/values/colors.xml
@@ -1,46 +1,48 @@
 <?xml version="1.0" encoding="utf-8"?>
 
-<!--version 4-->
 <resources>
-    <!-- Paleta principal -->
-    <color name="primary_color">#D12222</color> <!-- Azul moderno -->
+    <!-- Paleta Principal -->
+    <color name="primary_color">#D12222</color> <!-- Rojo principal -->
     <color name="secondary_color">#50C878</color> <!-- Verde esmeralda -->
-    <color name="tertiary_color">#9EFFD700</color> <!-- Dorado brillante -->
+    <color name="tertiary_color">#FFD700</color> <!-- Dorado brillante -->
     <color name="quaternary_color">#34495E</color> <!-- Azul grisáceo -->
 
-    <!-- Colores de fondo y texto -->
+    <!-- Colores de Fondo -->
     <color name="background_color">#F4F9FF</color> <!-- Fondo claro azulado -->
     <color name="background_dark">#1C2833</color> <!-- Fondo oscuro profundo -->
+    <color name="background_light">#ECF0F1</color> <!-- Fondo claro -->
+
+    <!-- Colores de Texto -->
     <color name="primary_text">#2C3E50</color> <!-- Texto principal -->
     <color name="secondary_text">#5D6D7E</color> <!-- Texto secundario -->
-    <color name="button_text">#FFFFFF</color> <!-- Texto en botones -->
+    <color name="tertiary_text">#9E9E9E</color> <!-- Texto terciario -->
 
-    <!-- Colores de botones -->
-    <color name="button_background_primary">#4A90E2</color> <!-- Botón principal -->
-    <color name="button_background_secondary">#50C878</color> <!-- Botón secundario -->
-    <color name="button_background_tertiary">#FFD700</color> <!-- Botón terciario -->
-    <color name="button_background_quaternary">#34495E</color> <!-- Botón cuarto -->
+    <!-- Colores de Botones -->
+    <color name="button_primary">#D12222</color> <!-- Botón principal -->
+    <color name="button_secondary">#50C878</color> <!-- Botón secundario -->
+    <color name="button_tertiary">#FFD700</color> <!-- Botón terciario -->
+    <color name="button_delete">#F44336</color> <!-- Botón eliminar -->
 
-    <!-- Colores de accesibilidad -->
+    <!-- Colores de Fondo de Botones -->
+    <color name="button_background_primary">#E74C3C</color> <!-- Fondo botón principal -->
+    <color name="button_background_secondary">#50C878</color> <!-- Fondo botón secundario -->
+    <color name="button_background_tertiary">#FFD700</color> <!-- Fondo botón terciario -->
+
+    <!-- Colores de Accesibilidad -->
     <color name="accessible_green">#27AE60</color> <!-- Verde accesible -->
     <color name="accessible_red">#E74C3C</color> <!-- Rojo accesible -->
 
-    <!-- Otros colores útiles -->
+    <!-- Colores Extras -->
     <color name="white">#FFFFFF</color>
     <color name="black">#000000</color>
     <color name="gray">#BDC3C7</color> <!-- Gris claro -->
-    <color name="warning_color">#F39C12</color> <!-- Amarillo advertencia -->
-
-    <!-- Colores claros para fondos -->
-    <color name="background_light">#ECF0F1</color>
-    <color name="tertiary_text">#9E9E9E</color> <!-- Gris más claro -->
-    <color name="button_delete">#F44336</color>
-    <color name="button_primary">#D12222</color> <!-- Azul moderno -->
-
-    <color name="primary_button">#D12222</color> <!-- Color para botones principales -->
-    <color name="link_text">#03A9F4</color>
+    <color name="link_text">#03A9F4</color> <!-- Texto de enlace -->
+    <color name="button_red">#E57373</color>
+    <color name="button_green">#81C784</color>
+    <color name="button_yellow">#FFF176</color>
+    <color name="button_blue">#64B5F6</color>
+    <color name="button_text">#FFFFFF</color>
 </resources>
-
 
 <!--version 3-->
 <!--<resources>-->
