+++ app/src/main/res/values/dimens.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <dimen name="guideline_percent">0.25</dimen>
+    <dimen name="title_text_size">28sp</dimen>
+    <dimen name="button_width">300dp</dimen>
+    <dimen name="button_height">60dp</dimen>
+    <dimen name="button_text_size">16sp</dimen>
+    <dimen name="button_margin_top">20dp</dimen>
+</resources>
