+++ app/src/main/res/values/themes.xml
@@ -12,8 +12,12 @@
         <item name="android:textColorPrimary">@color/primary_text</item>
         <item name="android:textColorSecondary">@color/secondary_text</item>
 
+        <!-- Fuente personalizada -->
+        <item name="fontFamily">@font/roboto</item>
+
         <!-- Colores para botones -->
-        <item name="buttonStyle">@style/ButtonStyle</item>
+        <item name="buttonStyle">@style/ButtonStyles</item>
+        <item name="materialButtonStyle">@style/MaterialButtonStyle</item>
     </style>
 
     <!-- Tema Oscuro -->
@@ -28,7 +32,11 @@
         <item name="android:textColorPrimary">@color/white</item>
         <item name="android:textColorSecondary">@color/gray</item>
 
+        <!-- Fuente personalizada -->
+        <item name="fontFamily">@font/roboto</item>
+
         <!-- Colores para botones -->
-        <item name="buttonStyle">@style/ButtonStyle</item>
+        <item name="buttonStyle">@style/ButtonStyles</item>
+        <item name="materialButtonStyle">@style/MaterialButtonStyle</item>
     </style>
 </resources>
