+++ gradle/libs.versions.toml
@@ -4,6 +4,8 @@ agp = "8.7.2"
 androidxJunit = "1.1.4"
 androidxLeanback = "1.2.0"
 appcompatVersion = "1.5.1"
+constraintlayout = "2.2.0"
+constraintlayoutVersion = "2.1.4"
 coreKtxVersion = "1.9.0"
 espressoCoreVersion = "3.5.0"
 googleMaterial = "1.7.0"
@@ -34,6 +36,8 @@ leanback = "1.0.0"
 [libraries]
 androidx-activity-compose-v161 = { module = "androidx.activity:activity-compose", version.ref = "activityComposeVersion" }
 androidx-appcompat-v151 = { module = "androidx.appcompat:appcompat", version.ref = "appcompatVersion" }
+androidx-constraintlayout = { module = "androidx.constraintlayout:constraintlayout", version.ref = "constraintlayout" }
+androidx-constraintlayout-v214 = { module = "androidx.constraintlayout:constraintlayout", version.ref = "constraintlayoutVersion" }
 androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "coreKtx" }
 androidx-core-ktx-v190 = { module = "androidx.core:core-ktx", version.ref = "coreKtxVersion" }
 androidx-espresso-core-v350 = { module = "androidx.test.espresso:espresso-core", version.ref = "espressoCoreVersion" }
